Extract common schema test utilities
